backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.2
pydantic-settings==2.3.0
firebase-admin==6.5.0
google-cloud-aiplatform==1.54.0
googlemaps==4.10.0
httpx==0.27.0
